Adding wizard modal service
The way angular-bootstrap modal work is a bit strange.
This is a problem because it requires us to define controller A for open.
And it also forces controller B to define close and dismiss.
The result is a bunch of controllers just to open a modal dialog.

This service aims to simplify that by centralizing open, close, and dismiss.
It also provides some default options that you must override via init.

Link to documentation:
https://angular-ui.github.io/bootstrap/#/modal

Change-Id: I730c7c34c69967dc74e4cb7827aec89a0c0d2b75
Partiall-Implements: blueprint angularize-identity-tables